feat(phase4): math/bits test suite validates in C# (package #14)
math/bits' own Go test suite now passes end-to-end through the -tests pipeline and matches
`go test -json`:

    Validated 26 tests against go test (0 skipped identically on both sides,
    79 disclosed-unsupported declarations excluded)

26/26, with NO disclosed-divergence rows -- nothing here needed the signature-pinned manifest; the
79 exclusions are the universally-deferred example/benchmark declarations (Phase 4D), the same
completion shape as the preceding packages. Reproduce with:

    go2cs -tests -test-action all "$(go env GOROOT)/src/math/bits" src/go-src-converted/math/bits

Per the validated-package commit policy the converted test sources are committed beside the
production code so the passing suite is visible and reviewable -- bits_test.cs, example_test.cs,
example_math_test.cs, export_test.cs, package_test_info.cs, go2cs_test_host.cs, and
math.bits.tests.csproj. The pipeline's regenerated inputs/outputs stay git-ignored.

Two hand-owned/refresh pieces make it run:

  - runtime is refreshed to current converter output. bits pulls runtime's overflowError and
    divideError by //go:linkname, and that forwarding property needs them PUBLIC (the linkname
    handle publicization) plus the errorString->ΔError adapter for `err.(runtime.Error)` to
    resolve; HEAD's committed runtime still had `internal static error divideError`. The
    production math.bits.csproj correspondingly gains the runtime ProjectReference -- which the
    converter already emits for a linkname pull, so this is adopting regenerated output (with the
    standard core\ -> go-src-converted\ relocation), not a hand edit.

  - runtime/panicvalues_impl.cs, a new hand-owned bridge with no Go counterpart, registers
    runtime's own divideError into golib's IntegerDivideByZeroValue hook via [ModuleInitializer].
    Div32 divides WITHOUT an explicit zero guard, so its panic comes from the CLR's hardware trap;
    the bridge makes that implicit panic carry the same value an explicit panicdivide() would, so
    TestDiv32PanicZero's `err.(runtime.Error)` assertion resolves. It carries
    [module: GoManualConversion] to state the ownership explicitly.

The other two blockers standing against this package are resolved rather than worked around: the
reflect .Clone CS1061 corpus blocker is fixed at the converter root in its own commit, and the
reported runtime efaceOf static-init crash does NOT reproduce in a correct configuration -- the
test host starts and runs 27/27 both through the pipeline and via a direct `dotnet run` of
math.bits.tests.csproj. That sighting was an artifact of an earlier throwaway overlay-only unblock
(a runtime built without the proper reconvert/overlay and without the runtime ProjectReference); no
converter or golib change was needed for it.

Scope note: go-src-converted is regenerable, so this commit deliberately banks ONLY math/bits and
the top-level runtime package rather than the full 302-package reconvert used to measure the work.
The trimmed tree was re-verified after the revert -- bits still validates 26/26.

+++ src/go-src-converted/runtime/sema.cs
@@ -320,7 +320,7 @@ internal static void queue(this ж<semaRoot> Ꮡroot, ж<uint32> Ꮡaddr, ж<sud
                     s.waittail = t;
                 }
                 s.waiters = t.Value.waiters;
-                if (s.waiters + 1 != 0) {
+                if ((uint16)(s.waiters + 1) != 0) {
                     s.waiters++;
                 }
                 t.Value.parent = default!;
@@ -336,7 +336,7 @@ internal static void queue(this ж<semaRoot> Ꮡroot, ж<uint32> Ꮡaddr, ж<sud
                 }
                 t.Value.waittail = Ꮡs;
                 s.waitlink = default!;
-                if ((~t).waiters + 1 != 0) {
+                if ((uint16)((~t).waiters + 1) != 0) {
                     t.Value.waiters++;
                 }
             }
